added gas estimation, tested on localhost.  Need to create guide on how to set this up, new testrpc is not that friendly

+++ src/contracts/Database.sol
@@ -1,5 +1,6 @@
 pragma solidity ^0.4.18;
 
+// TODO: Write changeContractManager() function
 contract Database { 
 
  
@@ -11,7 +12,7 @@ contract Database {
     mapping(bytes32 => bytes) public bytesStorage;
     mapping(bytes32 => bytes32) public bytes32Storage; 
     mapping(bytes32 => bool) public boolStorage;
-    mapping(bytes32 => int256) public intStorage;
+    mapping(bytes32 => int) public intStorage;
 
 
   // This is the first contract to get deployed. Will store all data + references to other contracts
